seperate StackUri and Stack in the inheritance tree
