+++ src/Image.php
@@ -12,7 +12,6 @@ use Rokka\Client\Core\SourceImage;
 use Rokka\Client\Core\SourceImageCollection;
 use Rokka\Client\Core\Stack;
 use Rokka\Client\Core\StackCollection;
-use Rokka\Client\Core\StackUri;
 
 /**
  * Image client for the rokka.io service.
@@ -370,9 +369,6 @@ class Image extends Base
         }
         if (empty($stack->getOrganization())) {
             $stack->setOrganization($this->defaultOrganization);
-        }
-        if ($stack instanceof StackUri) {
-            throw new \LogicException('Stack is a StackUrl object, should be a Stack object');
         }
 
         $queryString = [];
